Smart protocol tests reorganized to handle both local and S3 servers. Pull tests fully migrated. Push prepared, but not enabled for local FS yet.
