+++ kamu-adapter-http/tests/harness/mod.rs
@@ -16,6 +16,15 @@ pub use client_side_harness::*;
 mod server_side_harness;
 pub use server_side_harness::*;
 
+mod server_side_s3_harness;
+pub use server_side_s3_harness::*;
+
+mod server_side_local_fs_harness;
+pub use server_side_local_fs_harness::*;
+
+mod test_api_server;
+pub use test_api_server::*;
+
 macro_rules! await_client_server_flow {
     ($api_server_handle: expr, $client_handle: expr) => {
         tokio::select! {
