+++ kamu-adapter-http/tests/tests/tests_push/mod.rs
@@ -7,6 +7,5 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-mod test_routing;
-mod tests_pull;
-mod tests_push;
+mod test_smart_push_s3;
+mod test_smart_push_shared;
