+++ kamu-adapter-http/tests/tests/tests_push/test_smart_push_shared.rs
@@ -10,10 +10,10 @@
 use std::str::FromStr;
 
 use kamu::{
-    domain::*,
+    domain::{CommitOpts, DatasetExt, DatasetRepositoryExt, SyncResult},
     testing::{DatasetTestHelper, MetadataFactory},
 };
-use opendatafabric::*;
+use opendatafabric::{DatasetKind, DatasetRefLocal, DatasetRefRemote, MetadataEvent};
 
 use crate::harness::{
     await_client_server_flow, copy_folder_recursively, ClientSideHarness, ServerSideHarness,
@@ -21,10 +21,7 @@ use crate::harness::{
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-#[test_log::test(tokio::test)]
-#[cfg_attr(feature = "skip_docker_tests", ignore)]
-async fn test_smart_push_new_dataset() {
-    let server_harness = ServerSideHarness::new().await;
+pub async fn test_smart_push_new_dataset<T: ServerSideHarness>(server_harness: T) {
     let server_dataset_layout = server_harness.dataset_layout("foo");
 
     let client_harness = ClientSideHarness::new();
@@ -70,10 +67,7 @@ async fn test_smart_push_new_dataset() {
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-#[test_log::test(tokio::test)]
-#[cfg_attr(feature = "skip_docker_tests", ignore)]
-async fn test_smart_push_existing_up_to_date_dataset() {
-    let server_harness = ServerSideHarness::new().await;
+pub async fn test_smart_push_existing_up_to_date_dataset<T: ServerSideHarness>(server_harness: T) {
     let server_dataset_layout = server_harness.dataset_layout("foo");
 
     let client_harness = ClientSideHarness::new();
@@ -119,10 +113,7 @@ async fn test_smart_push_existing_up_to_date_dataset() {
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-#[test_log::test(tokio::test)]
-#[cfg_attr(feature = "skip_docker_tests", ignore)]
-async fn test_smart_push_existing_evolved_dataset() {
-    let server_harness = ServerSideHarness::new().await;
+pub async fn test_smart_push_existing_evolved_dataset<T: ServerSideHarness>(server_harness: T) {
     let server_dataset_layout = server_harness.dataset_layout("foo");
 
     let client_harness = ClientSideHarness::new();
@@ -190,10 +181,9 @@ async fn test_smart_push_existing_evolved_dataset() {
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-#[test_log::test(tokio::test)]
-#[cfg_attr(feature = "skip_docker_tests", ignore)]
-async fn test_smart_push_existing_dataset_fails_as_server_advanced() {
-    let server_harness = ServerSideHarness::new().await;
+pub async fn test_smart_push_existing_dataset_fails_as_server_advanced<T: ServerSideHarness>(
+    server_harness: T,
+) {
     let server_dataset_layout = server_harness.dataset_layout("foo");
     let server_repo = server_harness.dataset_repository();
